Add verbose flag for netsim cli
When verbose flag is not set only a minimal amount of information is
printed by netsim.
For example move, capture, reset, and radio commands will not output
anything when the command is successful.
devices command will only print a single line per device omitting
attributes that are of default value.

With verbose flag set netsim outputs the same information as before with
all the useful information in a slightly more concise formatting.

Bug:267375552

Test: ninja -C objs netsim
Test: objs/netsim devices
Test: objs/netsim -v devices
Change-Id: Ia3107d52a3f6275bf12b65803a9a170b92c0a886