Netsim python gRPC library skeleton
This contains the skeleton for a python gRPC library that
can be used by the emulator pytest framework to interact
with netsimd.

This will allow the pytest framework to make gRPC calls
with netsimd.

Currently it relies on emulator's prebuilt python interpreter and local
python packages under external/adt-infra/devpi/repo/simple

Bug: 284485910
Change-Id: Idb44f3fac62fa0b4e6fd3d4081621819c3ddea1c