Call close on transport to make sure Device::close is called.
When the bt facade gets a Remove(deviceId) ask the transport to call
the close-callback. This callback will invoke the HciDevice::Close()
which eventually sends disconnects to Acl handles and calls TestModel
close.

If facade Remove() is not being called in a timely manner then gRPC
service needs to be configured with keep-alive timeouts.

This is related to Avatar flakes.

Bug: 289280004

Test: m netsimd; netsimd
Change-Id: I1196c8a5fd68ffe278ed43472fecff5ed817989a