netsimd: For cuttlefish, do not autoshutdown immediately if -s argument is missing
If one builds for emulator and then runs without `-s` argument, the netsimd automatically shuts down after 15 seconds (which can be disabled with --no-shutdown). However if one builds for cuttlefish and runs without the `-s` argument (interactive) it shuts down immediately. Since logging by default is to a log output file, the error may be missed and this behavior may be confusing. This CL fixes this.

Test: atest libnetsim_daemon_tests
Test: m netsimd && ./out/host/linux-x86/bin/netsimd (aosp main)
Test: ninja -C objs netsimd && ./objs/netsimd (emu-master-dev)
Bug: 316365102
Bug: 316395279
Change-Id: I76ac0e04f2f6f7f030df82d5a305313f22372bd5