Hostapd - Clean up C args and streamline test
Remove unnecessary null termination in argv passed to C hostapd.

5s timeout for channel list update in hostapd is purposefully increased
from 1s previously so opted to keep as is. hostapd config field
country_code is optional but without it the beacon frames are not sent.

Updated and streamlined the integration test to share one instance of
hostapd to reduce overhead.

Bug: 377779061
Bug: 377778880
Test: scripts/build_tools.py --task runtest
Change-Id: Ib06431066f92b3effe7ccd39bf2d78c0cc28ab41