Run strict default Android lints in cmake build
Updates to run same set of Android default lints as Soong build.
Exception is `unnecessary_fallible_conversions` which requires clippy
1.75+.

RunTest now runs Clippy twice. First time it runs with default Rust
lints on Android which fails on any warning or error. Second time it
runs additional lints (which is just cognitive-complexity for now) and
does not fail the build on warnings.

Includes lots of lint fixes for netsim codebase.

Bug: 399147716
Bug: 402515839
Test: scripts/build_tools.py --task configure compile runtest
Change-Id: I7c5f968b0f54ca1c11dfe9e61b0a405564602876